+++ packages/pigeon/lib/src/generator_tools.dart
@@ -16,7 +16,7 @@ import 'generator.dart';
 /// The current version of pigeon.
 ///
 /// This must match the version in pubspec.yaml.
-const String pigeonVersion = '29.0.2';
+const String pigeonVersion = '29.0.3';
 
 /// Default plugin package name.
 const String defaultPluginPackageName = 'dev.flutter.pigeon';