+++ packages/vector_graphics_codec/test/validate_skills_test.dart
@@ -1,0 +1,34 @@
+// Copyright 2013 The Flutter Authors
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+@TestOn('vm')
+library;
+
+import 'dart:async';
+
+import 'package:logging/logging.dart';
+import 'package:skills_lint/skills_lint.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('Validate Skills', () async {
+    final Level oldLevel = Logger.root.level;
+    Logger.root.level = Level.ALL;
+    final StreamSubscription<LogRecord> subscription = Logger.root.onRecord.listen((
+      LogRecord record,
+    ) {
+      // ignore: avoid_print
+      print(record.message);
+    });
+
+    try {
+      final Configuration config = await ConfigParser.loadConfig();
+      final bool isValid = await validateSkills(config: config);
+      expect(isValid, isTrue, reason: 'Skills validation failed. See above for details.');
+    } finally {
+      Logger.root.level = oldLevel;
+      await subscription.cancel();
+    }
+  });
+}