+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.12.3
+
+* Adds agent skill for package setup and usage.
+
 ## 2.12.2
 
 * Fixes a [bug](https://github.com/flutter/flutter/issues/132934) where videos with a pixel aspect ratio other than 1.0 (anamorphic content) reported their coded size instead of their display size, causing them to be rendered stretched.